added more store state and  actions
