+++ src/views/SeriesList.vue
@@ -10,7 +10,7 @@
     <span v-if="!SearchedShow.length">
       <v-row no-gutters>
         <v-expansion-panels v-model="panel" variant="accordion">
-          <v-expansion-panel class="mt-4" v-for="(item, i) in differentGenres" :key="i">
+          <v-expansion-panel class="mt-4" v-for="(item, i) in store.differentGenres" :key="i">
             <v-expansion-panel-title class="exp-header-background" height="150">
               <h2 class="white--text">{{ item.label }}</h2>
               <template v-slot:actions>
@@ -30,11 +30,11 @@
     </span>
     <v-row no-gutters class="mt-4" v-else>
       <v-row v-if="!store.loading &&
-          searchedshowsData &&
-          searchedshowsData[0] &&
-          searchedshowsData[0].length
+          store.searchedshowsData &&
+          store.searchedshowsData[0] &&
+          store.searchedshowsData[0].length
           ">
-        <v-col v-for="item in searchedshowsData[0]" :key="item.id" xs="4" sm="4" md="2">
+        <v-col v-for="item in store.searchedshowsData[0]" :key="item.id" xs="4" sm="4" md="2">
           <show-card :showInfo="item.show" @card-click="cardClickedEvent" />
         </v-col>
       </v-row>
@@ -59,34 +59,12 @@ import { useTvStore } from '@/stores/tvStore'
 // import { debounceSearch } from '@/utils/debounce.js'
 import { useRouter } from 'vue-router'
 import ShowCard from "@/components/ShowCard.vue";
-import SeriesService from "@/seriesService/tv-service.js";
 
 
 const store = useTvStore()
 const router = useRouter()
-const SeriesService1 = new SeriesService();
-
-const differentGenres = ref([
-  { label: "War", filterdShows: [] },
-  { label: "Comedy", filterdShows: [] },
-  { label: "Thriller", filterdShows: [] },
-  { label: "Adventure", filterdShows: [] },
-  { label: "Action", filterdShows: [] },
-  { label: "Horror", filterdShows: [] },
-  { label: "Drama", filterdShows: [] },
-  { label: "Family", filterdShows: [] },
-  { label: "Romance", filterdShows: [] },
-  { label: "Supernatural", filterdShows: [] },
-  { label: "Science-Fiction", filterdShows: [] },
-  { label: "Medical", filterdShows: [] },
-  { label: "Crime", filterdShows: [] },
-  { label: "Legal", filterdShows: [] },
-  { label: "Fantasy", filterdShows: [] },
-])
 const panel = ref(0);
-const showsData = ref([]);
 const SearchedShow = ref("");
-const searchedshowsData = ref([]);
 
 function cardClickedEvent(v) {
   router.push({ path: `/SeriesDetails/${v}` });
@@ -96,62 +74,14 @@ function clearMessage() {
   SearchedShow.value = "";
 }
 
-/* get api call to fetch the list of tv series */
-
-function getTvSeriesData() {
-  showsData.value = [];
-  store.startLoading();
-  SeriesService1.seriesList()
-    .then((response) => {
-      showsData.value = response.data.sort((a, b) => {
-        return b.rating.average - a.rating.average;
-      });
-    })
-    .catch(() => {
-      showsData.value = [];
-    })
-    .finally(() => {
-      store.stopLoading();
-      getDiffGenresShows();
-    });
-}
-
-/* filters for shows on different genres */
-
-function getDiffGenresShows() {
-  for (let i = 0; i < differentGenres.value.length; i++)
-    differentGenres.value[i].filterdShows = showsData.value.filter((show) =>
-      show.genres.includes(differentGenres.value[i].label)
-    );
-}
-
-
-/* api call on search */
-
-function searchedShowDataFn(val) {
-  searchedshowsData.value = [];
-  store.startLoading();
-  SeriesService1.serachList(val)
-    .then((response) => {
-      searchedshowsData.value.push(response.data);
-    })
-    .catch(() => {
-      searchedshowsData.value = [];
-    })
-    .finally(() => {
-      store.stopLoading();
-    });
-}
-
-
 watch(SearchedShow, (srch) => {
-  if (srch.length)(searchedShowDataFn(srch));
+  if (srch.length)(store.searchedShowDataFn(srch));
   else {
-    searchedshowsData.value = [];
+    store.searchedshowsData = [];
   }
 })
-onMounted(()=>store.setTvSeriesName(''));
-getTvSeriesData();
+store.tvSeriesName='';
+store.getTvSeriesData();
 </script>
 <style scoped>
 .exp-header-background {
